Signup: Move site topic form to a separate component
