+++ client/signup/steps/site-topic/index.jsx
@@ -11,29 +11,13 @@ import { connect } from 'react-redux';
 /**
  * Internal dependencies
  */
-import Button from 'components/button';
-import StepWrapper from 'signup/step-wrapper';
-import FormFieldset from 'components/forms/form-fieldset';
-import SiteVerticalsSuggestionSearch, {
-	isVerticalSearchPending,
-} from 'components/site-verticals-suggestion-search';
-import { submitSiteVertical, setSiteVertical } from 'state/signup/steps/site-vertical/actions';
-import {
-	getSiteVerticalName,
-	getSiteVerticalSlug,
-	getSiteVerticalIsUserInput,
-	getSiteVerticalId,
-	getSiteVerticalParentId,
-} from 'state/signup/steps/site-vertical/selectors';
-import { getSiteType } from 'state/signup/steps/site-type/selectors';
-import { recordTracksEvent } from 'state/analytics/actions';
 import SignupActions from 'lib/signup/actions';
+import SiteTopicForm from './form';
+import StepWrapper from 'signup/step-wrapper';
+import { getSiteType } from 'state/signup/steps/site-type/selectors';
 import { getSiteTypePropertyValue } from 'lib/signup/site-type';
-
-/**
- * Style dependencies
- */
-import './style.scss';
+import { getSiteVerticalIsUserInput } from 'state/signup/steps/site-vertical/selectors';
+import { submitSiteVertical } from 'state/signup/steps/site-vertical/actions';
 
 class SiteTopicStep extends Component {
 	static propTypes = {
@@ -44,15 +28,11 @@ class SiteTopicStep extends Component {
 		submitSiteTopic: PropTypes.func.isRequired,
 		signupProgress: PropTypes.array,
 		stepName: PropTypes.string,
-		siteTopic: PropTypes.string,
-		siteSlug: PropTypes.string,
 		siteType: PropTypes.string,
 		translate: PropTypes.func.isRequired,
 	};
 
 	static defaultProps = {
-		siteTopic: '',
-		siteSlug: '',
 		isUserInput: true,
 	};
 
@@ -60,56 +40,6 @@ class SiteTopicStep extends Component {
 		SignupActions.saveSignupStep( {
 			stepName: this.props.stepName,
 		} );
-	}
-
-	onSiteTopicChange = verticalData => {
-		this.props.setSiteVertical( {
-			isUserInput: verticalData.isUserInputVertical,
-			name: verticalData.verticalName,
-			preview: verticalData.preview,
-			slug: verticalData.verticalSlug,
-			id: verticalData.verticalId,
-			parentId: verticalData.parent,
-		} );
-	};
-
-	onSubmit = event => {
-		event.preventDefault();
-		const {
-			isUserInput,
-			submitSiteTopic,
-			siteTopic,
-			siteSlug,
-			verticalId,
-			verticalParentId,
-		} = this.props;
-		submitSiteTopic( {
-			isUserInput,
-			name: siteTopic,
-			slug: siteSlug,
-			parentId: verticalParentId,
-			id: verticalId,
-		} );
-	};
-
-	renderContent() {
-		const { isButtonDisabled, siteTopic, translate } = this.props;
-		return (
-			<div className="site-topic__content">
-				<form onSubmit={ this.onSubmit }>
-					<FormFieldset>
-						<SiteVerticalsSuggestionSearch
-							onChange={ this.onSiteTopicChange }
-							initialValue={ siteTopic }
-							autoFocus={ true } // eslint-disable-line jsx-a11y/no-autofocus
-						/>
-						<Button type="submit" disabled={ isButtonDisabled } primary>
-							{ translate( 'Continue' ) }
-						</Button>
-					</FormFieldset>
-				</form>
-			</div>
-		);
 	}
 
 	getTextFromSiteType() {
@@ -138,7 +68,7 @@ class SiteTopicStep extends Component {
 					subHeaderText={ commonSubHeaderText }
 					fallbackSubHeaderText={ commonSubHeaderText }
 					signupProgress={ this.props.signupProgress }
-					stepContent={ this.renderContent() }
+					stepContent={ <SiteTopicForm submitForm={ this.props.submitSiteTopic } /> }
 					showSiteMockups={ this.props.showSiteMockups }
 				/>
 			</div>
@@ -147,16 +77,8 @@ class SiteTopicStep extends Component {
 }
 
 const mapDispatchToProps = ( dispatch, ownProps ) => ( {
-	submitSiteTopic: ( { id, isUserInput, name, parentId, slug } ) => {
+	submitSiteTopic: ( { isUserInput, name, slug } ) => {
 		const { flowName, goToNextStep, stepName } = ownProps;
-
-		dispatch(
-			recordTracksEvent( 'calypso_signup_actions_submit_site_topic', {
-				value: slug,
-				is_user_input_vertical: isUserInput,
-				parent_id: parentId || id,
-			} )
-		);
 
 		dispatch(
 			submitSiteVertical(
@@ -171,25 +93,14 @@ const mapDispatchToProps = ( dispatch, ownProps ) => ( {
 
 		goToNextStep( flowName );
 	},
-
-	setSiteVertical: verticalData => dispatch( setSiteVertical( verticalData ) ),
 } );
 
 export default localize(
 	connect(
-		state => {
-			const siteTopic = getSiteVerticalName( state );
-			const isButtonDisabled = ! siteTopic || isVerticalSearchPending();
-			return {
-				siteTopic,
-				siteSlug: getSiteVerticalSlug( state ),
-				siteType: getSiteType( state ),
-				isUserInput: getSiteVerticalIsUserInput( state ),
-				isButtonDisabled,
-				verticalId: getSiteVerticalId( state ),
-				verticalParentId: getSiteVerticalParentId( state ),
-			};
-		},
+		state => ( {
+			siteType: getSiteType( state ),
+			isUserInput: getSiteVerticalIsUserInput( state ),
+		} ),
 		mapDispatchToProps
 	)( SiteTopicStep )
 );
